requirements.txt
pytest
selenium
allure-pytest
psycopg2-binary